tests/wpt/web-platform-tests/css/css-sizing/parsing/min-width-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: getComputedStyle().minWidth</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-min-width">
<meta name="assert" content="min-width computed value is as specified, with <length-percentage> values computed">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("min-width", "min-content");
test_computed_value("min-width", "max-content");

test_computed_value("min-width", "10px");
test_computed_value("min-width", "20%");

test_computed_value('min-width', 'calc(10% + 40px)');
test_computed_value('min-width', 'calc(10px - 0.5em)', '0px');
test_computed_value('min-width', 'calc(10px + 0.5em)', '30px');

// fit-content not yet tested
</script>
</body>
</html>
