third_party/blink/web_tests/external/wpt/css/css-sizing/parsing/min-width-computed-expected.txt

This is a testharness.js-based test.
[PASS] Property min-width value 'min-content'
[PASS] Property min-width value 'max-content'
[PASS] Property min-width value '10px'
[PASS] Property min-width value '20%'
[PASS] Property min-width value 'calc(10% + 40px)'
[PASS] Property min-width value 'calc(10px - 0.5em)'
[PASS] Property min-width value 'calc(10px + 0.5em)'
[FAIL] Property min-width value 'fit-content(10px)'
  assert_true: 'fit-content(10px)' is a supported value for min-width. expected true got false
[FAIL] Property min-width value 'fit-content(20%)'
  assert_true: 'fit-content(20%)' is a supported value for min-width. expected true got false
[FAIL] Property min-width value 'fit-content(calc(10% + 40px))'
  assert_true: 'fit-content(calc(10% + 40px))' is a supported value for min-width. expected true got false
[FAIL] Property min-width value 'fit-content(calc(10px + 0.5em))'
  assert_true: 'fit-content(calc(10px + 0.5em))' is a supported value for min-width. expected true got false
Harness: the test ran to completion.
